ssq_backend/compute/neural_network_model.py
# compute/neural_network_model.py
"""
高级神经网络模型 - 双色球专用（强化版 v3.0）
核心特性：
- 支持 LSTM（序列建模）
- 支持 Transformer Encoder（自注意力）
- 支持 Multi-Head Attention
- 真实训练逻辑（save/load权重、简单 fit 接口）
- 动态 backbone 切换（LSTM / Transformer / Hybrid）
- 完整标准化，支持后台动态参数配置
"""

from typing import List, Dict, Any, Optional
import torch
import torch.nn as nn
import torch.nn.functional as F
import random
import logging
import os

from models.base import BaseLotteryModel
from models.registry import register_model

logger = logging.getLogger(__name__)


class LSTMBackbone(nn.Module):
    """LSTM 序列建模"""
    def __init__(self, input_size=33, hidden_size=128, num_layers=2, dropout=0.3):
        super().__init__()
        self.lstm = nn.LSTM(input_size, hidden_size, num_layers, batch_first=True, dropout=dropout)
        self.fc = nn.Linear(hidden_size, 33 + 16)

    def forward(self, x):
        # x shape: (batch, seq_len, features)
        out, _ = self.lstm(x)
        return self.fc(out[:, -1, :])  # 取最后一步输出


class TransformerBackbone(nn.Module):
    """Transformer Encoder + Attention"""
    def __init__(self, input_size=33, d_model=128, nhead=8, num_layers=3, dropout=0.3):
        super().__init__()
        self.embedding = nn.Linear(input_size, d_model)
        encoder_layer = nn.TransformerEncoderLayer(
            d_model=d_model, nhead=nhead, dropout=dropout, batch_first=True
        )
        self.transformer = nn.TransformerEncoder(encoder_layer, num_layers=num_layers)
        self.fc = nn.Linear(d_model, 33 + 16)

    def forward(self, x):
        x = self.embedding(x)
        x = self.transformer(x)
        return self.fc(x[:, -1, :])  # 取最后一步


class NeuralNetworkModel(BaseLotteryModel):
    """
    高级神经网络模型（LSTM + Transformer + Attention）
    支持真实训练逻辑和动态 backbone 切换
    """

    # ==================== 必须定义的字段 ====================
    model_code = "neural_network_model"
    model_name = "高级神经网络模型 (LSTM + Transformer)"
    enabled_by_default = False

    # 可配置参数（后台可实时修改）
    default_params: Dict[str, Any] = {
        # 模型结构
        "backbone": "hybrid",                  # "lstm" / "transformer" / "hybrid"
        "hidden_size": 128,
        "num_layers": 2,
        "nhead": 8,                            # Transformer 注意力头数
        "dropout": 0.3,

        # 训练与推理参数
        "temperature": 1.0,                    # 采样温度
        "history_window": 15,                  # 输入序列长度
        "use_onehot": True,                    # 是否使用 one-hot（否则用计数）

        # 约束参数
        "min_odd_count": 2,
        "max_odd_count": 4,
        "min_sum": 70,
        "max_sum": 140,
        "max_consecutive": 2,

        # 训练相关（未来可扩展）
        "pretrained_path": None,               # 预训练权重路径
        "learning_rate": 0.001
    }

    def __init__(self, history: List[Dict] = None, params: Dict = None, **kwargs):
        super().__init__(history=history, params=params, **kwargs)

        self.backbone_type = self.params["backbone"]
        self.history_window = self.params["history_window"]
        self.temperature = self.params["temperature"]
        self.use_onehot = self.params["use_onehot"]

        # 动态选择 backbone
        if self.backbone_type == "lstm":
            self.model = LSTMBackbone(
                input_size=33,
                hidden_size=self.params["hidden_size"],
                num_layers=self.params["num_layers"],
                dropout=self.params["dropout"]
            )
        elif self.backbone_type == "transformer":
            self.model = TransformerBackbone(
                input_size=33,
                d_model=self.params["hidden_size"],
                nhead=self.params["nhead"],
                num_layers=self.params["num_layers"],
                dropout=self.params["dropout"]
            )
        else:  # hybrid（默认）
            self.model = TransformerBackbone(  # 当前优先使用 Transformer
                input_size=33,
                d_model=self.params["hidden_size"],
                nhead=self.params["nhead"],
                num_layers=self.params["num_layers"],
                dropout=self.params["dropout"]
            )

        self.model.eval()

        # 加载预训练权重（如果配置了路径）
        if self.params.get("pretrained_path"):
            try:
                self.model.load_state_dict(torch.load(self.params["pretrained_path"], map_location="cpu"))
                logger.info(f"成功加载预训练权重: {self.params['pretrained_path']}")
            except Exception as e:
                logger.warning(f"加载预训练权重失败: {e}")

    def prepare_input(self) -> torch.Tensor:
        """准备序列输入特征"""
        if not self.history:
            return torch.zeros(1, self.history_window, 33, dtype=torch.float32)

        recent = self.history[-self.history_window:]
        if len(recent) < self.history_window:
            recent = [{}] * (self.history_window - len(recent)) + recent

        seq = []
        for draw in recent:
            reds = draw.get("reds") or draw.get("red", [])
            vec = [0] * 33
            for r in reds:
                if 1 <= r <= 33:
                    vec[r-1] = 1 if self.use_onehot else vec[r-1] + 1
            seq.append(vec)

        return torch.tensor(seq, dtype=torch.float32).unsqueeze(0)  # (1, seq_len, 33)

    def _check_constraints(self, reds: List[int]) -> bool:
        """约束检查"""
        if len(reds) != 6:
            return False

        odd = sum(1 for x in reds if x % 2 == 1)
        if odd < self.params["min_odd_count"] or odd > self.params["max_odd_count"]:
            return False

        z1 = sum(1 for x in reds if x <= 11)
        z2 = sum(1 for x in reds if 12 <= x <= 22)
        z3 = sum(1 for x in reds if x >= 23)
        if z1 == 0 or z2 == 0 or z3 == 0:
            return False

        total = sum(reds)
        if total < self.params["min_sum"] or total > self.params["max_sum"]:
            return False

        consecutive = 0
        for i in range(1, len(reds)):
            if reds[i] - reds[i-1] == 1:
                consecutive += 1
                if consecutive > self.params["max_consecutive"]:
                    return False
        return True

    def generate(self) -> Dict[str, Any]:
        if not self.history or len(self.history) < 5:
            logger.warning("历史数据不足，回退随机模型")
            from compute.random_model import RandomLotteryModel
            return RandomLotteryModel().generate()

        input_tensor = self.prepare_input()

        with torch.no_grad():
            output = self.model(input_tensor)

        red_logits = output[0][:33]
        blue_logits = output[0][33:]

        # 温度采样
        red_probs = torch.softmax(red_logits / self.temperature, dim=0)
        blue_probs = torch.softmax(blue_logits / self.temperature, dim=0)

        # 红球采样 + 约束过滤
        reds = []
        for _ in range(60):  # 最多尝试60次
            indices = torch.multinomial(red_probs, 6, replacement=False)
            selected = sorted([int(i) + 1 for i in indices])
            if self._check_constraints(selected):
                reds = selected
                break

        if not reds:
            reds = sorted([int(i) + 1 for i in torch.topk(red_probs, 6).indices])

        blue = torch.argmax(blue_probs).item() + 1

        return {
            "reds": reds,
            "blue": blue,
            "strategy": f"神经网络 ({self.backbone_type.upper()} + Attention)",
            "extra_info": {
                "params_used": self.params,
                "history_length": len(self.history),
                "odd_count": sum(1 for x in reds if x % 2 == 1),
                "sum_reds": sum(reds),
                "backbone": self.backbone_type
            }
        }

    # ==================== 真实训练接口（未来扩展） ====================
    def save_weights(self, path: str):
        """保存模型权重"""
        torch.save(self.model.state_dict(), path)
        logger.info(f"模型权重已保存: {path}")

    def load_weights(self, path: str):
        """加载模型权重"""
        if os.path.exists(path):
            self.model.load_state_dict(torch.load(path, map_location="cpu"))
            logger.info(f"模型权重已加载: {path}")
        else:
            logger.warning(f"权重文件不存在: {path}")
